% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_tar_target.R
\name{insert_tar_target}
\alias{insert_tar_target}
\title{Insert a tar_target() skeleton}
\usage{
insert_tar_target()
}
\value{
Invisibly returns `NULL`.
}
\description{
RStudio addin that inserts a `targets::tar_target()` skeleton at the current
cursor position.
}
\details{
NB: a comma is intentionally added at the end of the text in order to
facilitate intra-target list insertion which the package author finds is
the most common situation
}
\examples{
\dontrun{
# In RStudio, place the cursor in an R script and run:
insert_tar_target()

# Or use the registered RStudio addin:
# Addins > Insert tar_target

# The addin inserts:
# tar_target(
#    name = ,
#    command =
#  ),
}

}
