% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targets.utils-package.R
\docType{package}
\name{targets.utils-package}
\alias{targets.utils}
\alias{targets.utils-package}
\title{targets.utils}
\description{
Utilities for working with targets pipelines.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/geryan/targets.utils}
  \item \url{https://geryan.github.io/targets.utils/}
  \item Report bugs at \url{https://github.com/geryan/targets.utils/issues}
}

}
\author{
\strong{Maintainer}: Gerry Ryan \email{go@aw.ay} (\href{https://orcid.org/0000-0003-0183-7630}{ORCID}) [copyright holder]

}
\keyword{internal}
