% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl.R
\name{tl}
\alias{tl}
\title{Targets load globals and everything}
\usage{
tl()
}
\value{
Nothing
}
\description{
Runs `targets::tar_load_globals` then `targets::tar_load_everything` to save
you the drudgery of typing those thing out again and again and again and
again and again et cetera ad nauseum.
}
\examples{
\dontrun{
tl()
}
}
