% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tml.R
\name{tml}
\alias{tml}
\title{Target make and load}
\usage{
tml(x)
}
\arguments{
\item{x}{Object name to make and load}
}
\value{
Nothing
}
\description{
Convenience function to make and load an object when working interactively with a targets workflow. Calls `targets::tar_make(x)` then `targets::tar_load(x)`
}
\examples{
\dontrun{
tml(x)
}
}
\author{
Gerry Ryan
}
