% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tar_downstream.R
\name{tar_downstream}
\alias{tar_downstream}
\title{NOT WORKING PROPERLY YET}
\usage{
tar_downstream(target_name)
}
\arguments{
\item{target_name}{Unquoted target name.}
}
\value{
Character vector of downstream target names.
}
\description{
NOT WORKING PROPERLY YET
}
\examples{
\dontrun{tar_downstream(my_target)}

}
\keyword{internal}
